requirements.txt
# UNDLM PostFlow - Requirements
# Pipeline de post-production pour documentaire animé

# Core dependencies
requests>=2.28.0
pathlib2>=2.3.0
dataclasses>=0.8; python_version < "3.7"
pandas>=1.5.0
python-dotenv>=0.19.0
watchdog>=2.1.0
schedule>=1.1.0
pyyaml>=6.0

# Discord integration
discord-webhook>=1.3.0

# Frame.io integration with remote upload & Cloudflare Tunnel
httpx>=0.24.0
frameioclient>=2.0.0

# Advanced monitoring and optimization (v4.1.0)
aiohttp>=3.8.0
psutil>=5.9.0

# Google Sheets integration
gspread>=5.7.0
google-auth>=2.16.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0

# After Effects integration (via ExtendScript)
# Note: Communication via file system, no Python package needed

# Error handling and monitoring
flask>=2.0.0
flask-socketio>=5.0.0
python-socketio>=5.0.0
eventlet>=0.33.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: for enhanced logging
colorama>=0.4.6

# Code quality
black>=22.0.0
flake8>=5.0.0

# Optional: for data analysis
pandas>=1.5.0
numpy>=1.24.0

# Frame.io v4 integration dependencies
httpx>=0.27.0
PyJWT[crypto]>=2.8.0
cryptography>=41.0.0
python-dotenv>=1.0.0
